Chore: move issue_command match to handle_issue command
