+++ src/cli_parse/handle_cli.rs
@@ -1,15 +1,13 @@
 use crate::cli_in::comment_command::CommentCommand;
-use crate::cli_in::issue_command::IssueCommand;
 use crate::cli_in::read_cli::Args;
 use crate::cli_in::read_cli::CliCommand;
 use crate::cli_in::release_command::ReleaseCommand;
 use crate::cli_in::repo_command::RepoCommand;
-use crate::cli_out::print_in_cli::print_issues;
 use crate::cli_out::print_in_cli::print_release;
 use crate::cli_out::print_in_cli::print_repos;
 use crate::cli_out::print_in_cli::print_url;
+use crate::cli_parse::handle_commands::handle_issue::handle_issue_command;
 use crate::git_utils::comments;
-use crate::git_utils::issues;
 use crate::git_utils::releases;
 use crate::git_utils::repo_info::Repo;
 use crate::git_utils::repo_info::RepoInfo;
@@ -19,122 +17,9 @@ use std::process;
 
 pub async fn handle_cli_command(args: Args, github_client: Client) {
     match args.command {
-        CliCommand::Issue { subcommand } => match subcommand {
-            IssueCommand::List {
-                creator,
-                assignee,
-                state,
-                labels,
-                numb_of_page,
-                iss_on_page,
-            } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Current, None, None) {
-                    Ok(rep) => rep,
-                    Err(message) => {
-                        eprintln!("Error: {message}");
-                        process::exit(1);
-                    }
-                };
-
-                let list_issues = issues::get_list(
-                    &github_client,
-                    &repo_info,
-                    &creator,
-                    &assignee,
-                    &state.0,
-                    &labels,
-                    &numb_of_page,
-                    &iss_on_page,
-                )
-                .await;
-
-                print_issues(list_issues, state, numb_of_page);
-            }
-
-            IssueCommand::Create {
-                title,
-                body,
-                assignees,
-                labels,
-            } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Current, None, None) {
-                    Ok(rep) => rep,
-                    Err(message) => {
-                        eprintln!("Error: {message}");
-                        process::exit(1);
-                    }
-                };
-                let labels_list: Vec<String> = labels.split(",").map(|s| s.to_string()).collect();
-                let assignees_list: Vec<String> =
-                    assignees.split(",").map(|s| s.to_string()).collect();
-
-                let result = issues::create(
-                    &github_client,
-                    repo_info,
-                    &title,
-                    &body,
-                    &assignees_list,
-                    &labels_list,
-                )
-                .await;
-
-                println!("{result}");
-            }
-
-            IssueCommand::Close { number, comment } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Current, None, None) {
-                    Ok(rep) => rep,
-                    Err(message) => {
-                        eprintln!("Error: {message}");
-                        process::exit(1);
-                    }
-                };
-                let result = issues::close(&github_client, repo_info, &number, &comment).await;
-
-                println!("{result}");
-            }
-
-            IssueCommand::Update {
-                number,
-                title,
-                body,
-                assignees,
-                state,
-                labels,
-            } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Current, None, None) {
-                    Ok(rep) => rep,
-                    Err(message) => {
-                        eprintln!("Error: {message}");
-                        process::exit(1);
-                    }
-                };
-
-                let labels_list: Vec<String> = match labels {
-                    Some(l) => l.split(",").map(|s| s.to_string()).collect(),
-                    None => Vec::new(),
-                };
-                let assignees_list: Vec<String> = match assignees {
-                    Some(a) => a.split(",").map(|s| s.to_string()).collect(),
-                    None => Vec::new(),
-                };
-
-                let result = issues::update(
-                    &github_client,
-                    repo_info,
-                    &number,
-                    Some(title),
-                    Some(body),
-                    &assignees_list,
-                    &labels_list,
-                    &state.0,
-                )
-                .await;
-
-                println!("{result}");
-            }
-        },
-
+        CliCommand::Issue { subcommand } => {
+            handle_issue_command(github_client, subcommand).await;
+        }
         CliCommand::Comment { subcommand } => match subcommand {
             CommentCommand::Create { number, body } => {
                 let repo_info: RepoInfo = match RepoInfo::new(Repo::Current, None, None) {
@@ -273,7 +158,8 @@ pub async fn handle_cli_command(args: Args, github_client: Client) {
                 private,
                 include_all_branches,
             } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(name)) {
+                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(name))
+                {
                     Ok(rep) => rep,
                     Err(message) => {
                         eprintln!("Error: {message}");
@@ -303,7 +189,8 @@ pub async fn handle_cli_command(args: Args, github_client: Client) {
             }
 
             RepoCommand::CreateFork { org, name, owner } => {
-                let fork_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(name)) {
+                let fork_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(name))
+                {
                     Ok(rep) => rep,
                     Err(message) => {
                         eprintln!("Error: {message}");
@@ -346,7 +233,8 @@ pub async fn handle_cli_command(args: Args, github_client: Client) {
                 tag_name,
                 target_commitish,
             } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo)) {
+                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo))
+                {
                     Ok(rep) => rep,
                     Err(message) => {
                         eprintln!("Error: {message}");
@@ -371,7 +259,8 @@ pub async fn handle_cli_command(args: Args, github_client: Client) {
             }
 
             ReleaseCommand::GetLatest { owner, repo } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo)) {
+                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo))
+                {
                     Ok(rep) => rep,
                     Err(message) => {
                         eprintln!("Error: {message}");
@@ -385,7 +274,8 @@ pub async fn handle_cli_command(args: Args, github_client: Client) {
             }
 
             ReleaseCommand::GetByTag { owner, repo, tag } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo)) {
+                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo))
+                {
                     Ok(rep) => rep,
                     Err(message) => {
                         eprintln!("Error: {message}");
@@ -399,7 +289,8 @@ pub async fn handle_cli_command(args: Args, github_client: Client) {
             }
 
             ReleaseCommand::GetById { owner, repo, id } => {
-                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo)) {
+                let repo_info: RepoInfo = match RepoInfo::new(Repo::Input, Some(owner), Some(repo))
+                {
                     Ok(rep) => rep,
                     Err(message) => {
                         eprintln!("Error: {message}");
