+++ src/cli_parse/mod.rs
@@ -1,1 +1,2 @@
 pub mod handle_cli;
+mod handle_commands;
